library: firmware_update: fix on-the-fly install using manifest.txt
On-the-fly update using a manifest.txt file is not possible: the fragments of
the swu with the update must be already in the device file system when the
manifest update starts.

This commit ignores the value of on-the-fly setting and proceeds as if
on-the-fly is disabled, if the firmware update target is 'manifest.txt'.

https://onedigi.atlassian.net/browse/DEL-8399

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>