library: cc_init: retry the connection after downloading the client certificate
If 'enable_reconnection' was not enabled in the configuration and the client
certificate must be downloaded (because it is the first connection or it was
reset in DRM), the connector was not trying to reconnect.

A reconnect is always:
 * required after downloading client certificate, independently of the
   'enable_reconnection' setting value
 * executed after downloading the certificate without waiting the timeout
   configured in 'reconnect_time' setting

The rest of the reconnections must follow what is configured in settings
'enable_reconnection' and 'reconnect_time'.

This commit reimplements commit 57f47a0c54a6573861d572af740a93c48402684b.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>