library: fix stop of connection with the cloud
The stop sequence of cloud connector includes 2 phases:

1. Stop transports
2. Stop CCAPI

Sometimes, during the shutdown, the cloud connector state machine does not work
properly and the process gets blocked waiting for a status change that is not
going to happen. This results the connector does not shutdown, the code blocks,
and a manual kill of the process is required.
It seems that if the first phase did not proprely complete before the second
one starts, the state machine does not advance properly.

We added a one second sleep between these 2 steps to workaround this issue.
We tested a loop of more than 600 shutdowns successfully.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>