library: cccs: if socket is closed do not try to send an error
We saw this on a CC6 running the 'cccs-gs-demo' application: the first time the
socket recieve data to register data requests from the application, the 'select'
signals there is data to read, but the 'read' operation returns 0 bytes, which
means the socket is closed. The code considers this an error and tries to send
the error message using the socket and a segmentation fault occurs.

This commit tries to avoid this situation, so if 0 bytes are read no message
is sent over the socket.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>