library: firmware: otf: fix sync between read data from DRM and on-the-fly write threads
After increasing the size of the chunk for the firmware update (see commits
below), the download process is faster and may corrupt the data while being
written by swupdate.

To avoid this, this commit includes a synchronization mechanism between the
reception data thread (function 'firmware_data_cb') and the write data thread
(function 'otf_read_image_cb'):

  * Reception data thread must wait until swupdate finishes with the previous
    data chunk before providing new received data.
  * Swupdate write thread must wait until new data chunk is received from DRM.

This commit also synchronizes the end of the swupdate process (function
'otf_end_cb') if an error occurs before all data is received. In this situation,
the update process must end with an error.

See commits:
 * 8e100e94463b71f87d0eedbad470285813b8942c
   library: cc_api: add maximum firmware binary chunk size
 * 6832ff299ccbede0e28ca704db911a5be8ad04bb
   library: firmware: define 10K as the maximum received packet size

https://onedigi.atlassian.net/browse/DEL-8878

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>